Sprint 5

+++ Store/Basket.Host/Services/BasketService.cs
@@ -1,5 +1,7 @@
+using System.Net;
 using Basket.Host.Models;
 using Basket.Host.Services.Interfaces;
+using ExceptionHandler;
 
 namespace Basket.Host.Services;
 
@@ -21,7 +23,8 @@ public class BasketService: IBasketService
     
     public async Task AddItem(string userId, Item item)
     {
-        await _httpClient.SendAsync<Task, object>(
+        _logger.LogInformation($"*{GetType().Name}* adding {item.Quantity} item with id: {item.ItemId}");
+        var response = await _httpClient.SendAsync<Task, object>(
             $"{_catalogBaseUrl}/items/decrease", 
             HttpMethod.Put, new List<Item>()
             {
@@ -30,12 +33,17 @@ public class BasketService: IBasketService
                     ItemId = item.ItemId, Quantity = item.Quantity
                 }
             });
+        if (response.Status == TaskStatus.Faulted)
+        {
+            throw new IllegalArgumentException(response.Exception?.InnerException?.Message ?? "Unknown error");
+        }
         await _cacheService.AddOrUpdateAsync(userId, item.ItemId, item.Quantity);
     }
 
     public async Task RemoveItem(string userId, Item item)
     {
-        await _httpClient.SendAsync<Task, object>(
+        _logger.LogInformation($"*{GetType().Name}* request to remove {item.Quantity} items: {item.ItemId}");
+        var response = await _httpClient.SendAsync<Task, object>(
             $"{_catalogBaseUrl}/items/increase", 
             HttpMethod.Put, new List<Item>()
             {
@@ -44,53 +52,45 @@ public class BasketService: IBasketService
                     ItemId = item.ItemId, Quantity = item.Quantity
                 }
             });
+        if (response.Status == TaskStatus.Faulted)
+        {
+            throw new Exception("Unknown error");
+        }
         await _cacheService.RemoveOrUpdateAsync(userId, item.ItemId, item.Quantity);
     }
 
     public async Task<List<OrderItem>> GetItems(string userId)
     {
         var items = await GetBasketItems(userId);
+        _logger.LogInformation($"*{GetType().Name}* found {items.Count} items in the basket");
         List<OrderItem> orderItems = new List<OrderItem>();
         if (items.Count == 0)
         {
             return orderItems;
         } 
-        foreach (var i in items)
-        {
-            var catalogItem =  await _httpClient.SendAsync<ItemModel, object>(
-                $"{_catalogBaseUrl}/items/{i.ItemId}", 
-                HttpMethod.Get, null);
-           orderItems.Add(new OrderItem()
-           {
-               ItemId = i.ItemId,
-               BrandId = catalogItem.CatalogItem.ItemBrandId,
-               Price = catalogItem.CatalogItem.Price,
-               Quantity = i.Quantity,
-               Size = catalogItem.Size,
-               Name = catalogItem.CatalogItem.Name,
-               Image = catalogItem.CatalogItem.Image,
-               StockQuantity = catalogItem.Quantity
-           });
-        }
         
+        _logger.LogInformation($"*{GetType().Name}* items in the basket: {orderItems.ToString()}");
         return orderItems;
     }
 
     private async Task<List<Item>> GetBasketItems(string userId)
     {
         var items = await _cacheService.GetAsync(userId);
-        return items
+        var basketItems =  items
             .Select(entry => new Item
             {
                 ItemId = int.Parse(entry.Name), 
                 Quantity = int.Parse(entry.Value),
             })
             .ToList();
+        _logger.LogInformation($"*{GetType().Name}* total items in the basket: {basketItems.Count}");
+        return basketItems;
     }
 
     public async Task RemoveAll(string userId)
     {
         var items = await GetBasketItems(userId);
+        _logger.LogInformation($"*{GetType().Name}* total items to remove: {items.Count}");
         await _cacheService.RemoveAllAsync(userId);
         await _httpClient.SendAsync<Task, object>(
             $"{_catalogBaseUrl}/items/increase", 
@@ -100,10 +100,11 @@ public class BasketService: IBasketService
     public async Task<int> CheckoutBasket(string userId)
     {
         var items = await GetItems(userId);
-        
+        _logger.LogInformation($"*{GetType().Name}* total items to check-out: {items.Count}");
         var orderId = await _httpClient.SendAsync<int, object>(
             $"{_orderBaseUrl}/orders/{userId}", 
             HttpMethod.Post, items);
+        _logger.LogInformation($"*{GetType().Name}* check-out completed with order id: {orderId}");
         await RemoveAll(userId);
         return orderId;
     }
